FastAPI_DDD/pyproject.toml
[project]
name = "fastapi-ddd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aerich>=0.9.2",
    "aiomysql>=0.3.2",
    "fastapi>=0.124.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "tomlkit>=0.13.3",
    "tortoise-orm>=0.25.1",
    "uvicorn[standard]>=0.38.0",
    "jinja2>=3.1.0,<4.0.0",
    "markupsafe>=2.1.0,<3.0.0",
    "pyjwt>=2.7.0",
    "redis>=5.1.0",
    "fastapi-mail>=1.4.2",
    "psutil>=5.9.0",
]
